Generate Expo project
This is the (almost) unmodified output of running `expo init` in our
project directory. We changed the name to `hodor` (instead of the
default `hodor-expo`), as well as selecting the Typescript tabs
template. The generator placed all files within a subdirectory named
`hodor`, so we moved files up to the project root and deleted the now-
empty `hodor` directory.