Add and configure `expo-dev-client`
This provides better troubleshooting and debugging tools when running
internal Android and iOS builds. Note that we've added it to our normal
dependencies section (and not dev dependencies) since this would also be
used in preview builds published internally.

We've also imported this package in our `App.tsx` to provide better
error messages. See here for more info:
https://docs.expo.dev/development/installation/#add-better-error-handlers.