Use open-source wave icon instead of defaults
This just swaps out the existing default expo icon and splash images for
the open-source wave icon. The others have been deleted and we've run
`npx expo-optimize` to update our `assets.json` file (no modifications
were made to the image as it was already fully optimized). Running this
also automatically created a readme in the `.expo-shared` directory so
we're no longer confused about its existence (:

The image was retrieved from here:
https://openmoji.org/library/emoji-1F44B.