Define application ids
This gives us two application ids for our two channels. We define it in
the environment in `eas.json`, which is then accessible in our Expo
application config.

Note: this means that in _local_ development, `APPLICATION_ID` is never
defined! This doesn't matter though, as the application id is only
important when we're producing builds that get installed directly on
devices (and not via Expo Go).